pyproject.toml
[project]
name = "ai-resume"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "gradio>=5.49.1",
    "openai>=2.7.2",
    "pypdf>=6.2.0",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
]
